requirements.txt
pandas>=1.2.0
numpy>=1.19.0
matplotlib>=3.3.0
seaborn>=0.11.0
scikit-learn>=0.24.0
lightgbm>=3.2.0
xgboost>=1.4.0
catboost>=0.24
tensorflow>=2.6.0
statsmodels>=0.12.0
jupyter>=1.0.0
notebook>=6.4.0
tqdm>=4.62.0
pyyaml>=5.4.1
python-dotenv>=0.19.0 